keeping stream_deepspeech file unchanged and pulled the work over to the mapping file
